feat(TrackSegment): Pass 3 — typed scene root and meshes
Migrate index.jsx, TrackSegmentMeshes.jsx, and PondFog.jsx to TypeScript with
explicit TrackSegmentProps / TrackSegmentMeshesProps (BiomeId-normalized, no
public any). Fix hemisphereLight skyColor → color for R3F typings. Add a
thin meander headless page-error harness for the migration gate.
